fix(desktop): collapse toolHint progress after final + e2e shared helpers (#184)
* fix(desktop): collapse toolHint progress after final + e2e shared helpers

Three fixes:

1. removeTransientTurnMessagesSinceLastUser: mark retained toolHint
   progress with collapsed=true so they render collapsed after cleanup.

2. sendCleanup: stop calling cleanupListeners() when the typewriter
   finishes. Another final may arrive in the same turn. Listeners are
   now torn down only on abort/error/timeout/new-session.

3. onFinal: merge cleanup+dedup into single setMessages. Skip appending
   collapsed tool-call group when streaming already rendered toolHint
   progress to avoid duplicate entries.

e2e: extract shared Electron launch/bridge/helper into electron-setup.ts,
eliminating ~70 lines of beforeAll duplication per spec file.

Fixes 3 smoke tests in issue-172-assistant-turn-collapse.spec.ts.


* fix(test): polyfill requestAnimationFrame in mock bridge to fix flaky smoke tests

The ChatConsole typewriter animation relies on requestAnimationFrame
for character-by-character rendering. In headless Chromium (Playwright),
idle pages throttle rAF to 1 fps or stop it entirely, causing
expect(...).toBeVisible() to timeout for long final responses.

Polyfill rAF with setTimeout(fn, 0) in the mock bridge script so the
typewriter completes instantly in smoke tests, making them
deterministic. Fixes the flaky issue-172-assistant-turn-collapse test.

---------
